requirements.txt
accelerate>=0.33.0
openai>=1.0.0
bitsandbytes>=0.43.1
datasets>=2.20.0
httpx>=0.27.0
huggingface_hub>=0.24.6
numpy>=1.26.0
pydantic>=2.8.2
python-dotenv>=1.0.1
scikit-learn>=1.5.1
tenacity>=8.5.0
tokenizers>=0.19.1
torch>=2.3.1
tqdm>=4.66.4
transformers>=4.44.0
trl>=0.9.6
peft>=0.11.1
unsloth>=2024.7.6
